restoring cyclone library
the original ready description described how cyclone could be set as a
regular lib. The -lib cyclone would call hammer and sickle objects plus
a small set of 12 objects with no alphanumerical names (old nettles.c
file).

Seems that in the Pd Extended days, the regular lib was split in
individual objects binaries for hammer/sickle but those 12 objects were
lost in this process.

I am now restoring the original cyclone library concept, to be included
in Pd’s startup, to load these 12 objects that were lost.

Recently, a “nettles” library was created to include these, but I
consider this a diversion from the original idea of having a cyclone
library, and I, for one, have never seen the word “nettles” and find it
hard, as a user, to relate it to cyclone…

This update successfully created a cyclone library that loads these 12
objects