removing "purr" stuff
Removing this is a documentation for purr data that was supposed to be a quick and temporary solution for including cyclone in purr data, mostly because [comment] hadn't been ported yet, to be removed when [comment] would finally be ported...

but cyclone wasn't included in purr data and it's been quite a while. The actual documentation suffered many edits and now it's just best to port [comment] to purr data and use one single documentation.

we're struggling to deal with [comment] in tcl tk, unfortunately we're not likely to port comment ourselves, as well as other objects in cyclone that need to be ported, see:https://git.purrdata.net/jwilkes/purr-data/issues/355